put wavefront stuff into a file that can be pulled into p1 or p2

+++ ws/shared/Helpers.h
@@ -59,14 +59,14 @@ namespace H{
     inline Eigen::Vector2d obstaclesClosePt(const amp::Environment2D& env, const Eigen::Vector2d& q);
 
     ///@brief continuous number to index
-    inline int numToIdx(double num, double num_min, double num_max, int len_arr)
-    {int idx = (num-num_min)/(num_max-num_min)*len_arr; 
-    if(idx>=len_arr) return len_arr-1;
-    else return idx;}
+    inline int numToIdx(double num, double num_min, double num_max, int len_arr){
+        int idx = (num-num_min)/(num_max-num_min)*len_arr; 
+        if(idx>=len_arr) return len_arr-1;
+        else return idx; }
 
     ///@brief index to continous number
     inline double idxToNum(int idx, int len_arr, double cont_min, double cont_max)
-    {return (double(idx)/len_arr*(cont_max-cont_min)) + cont_min + ((cont_max-cont_min)/len_arr);}
+    {return (double(idx)/len_arr*(cont_max-cont_min)) + cont_min + ((cont_max-cont_min)/double(len_arr));}
 
     ///@brief returns a random collision free point within the environment bounds
     inline Eigen::Vector2d randomValidSample(const amp::Problem2D& problem, int n=20);
@@ -456,3 +456,5 @@ inline std::vector<Eigen::Vector2d> H::getAllPgVerts(const amp::Environment2D& e
     }
     return all_pts;
 }
+
+
